Use cython to speed up computations
Separates the functions in maps.py into several files:
  maps.py - more generic, high-level stuff
  formats.py - how to write in asc, ply, stl
  projections.pyx - cython file with projection-related stuff

The generated projections.c file is included so that
users do not need to have cython installed. It is created
with: cython3 -a projections.pyx

The setup.py file is added so the module can be created with:
  python3 setup.py build_ext --inplace

This way we get a 2x speed boost. Yay.